requirements.txt
selenium
python-dotenv
requests